Rename sass/script/functions/meta.rb to sass/script/functions/introspection.rb and move some functions there from sass/script/functions/number.rb
